pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "email-validator>=2.2.0",
    "faiss-cpu>=1.10.0",
    "flask>=3.1.0",
    "flask-sqlalchemy>=3.1.1",
    "gunicorn>=23.0.0",
    "langchain>=0.3.23",
    "numpy>=2.2.4",
    "openai>=1.70.0",
    "psycopg2-binary>=2.9.10",
    "requests>=2.32.3",
    "sqlalchemy>=2.0.40",
    "werkzeug>=3.1.3",
]
